refactoring login tests for the other two specifications.
The idea is to be able to make them work in parallel except for the tests inside a single KarXXX which are supposed to be launched in parallel.

Also having split up the subtests makes it quite easy in intellij to find for example ctrl-N Kar136 will find at the same time
Kar136_test
Kar136_route
Kar136_actor
so simplifying the navigation and maintenance.