almost at the point of fully testing new additions from Chandan.
I did rollback the ts Option[DateTime] back to simple DateTime with default = now I don't see any valid reason for having it optional when it is actually always set up.
when copying it with .copy it can be easily changed to new now without any problems.

There is still some confusion on Karedos, being in the db as Double, but actually double is quite dangerous to keep currencies (we Italians did know very well this having Liras in the past which was similar to karedos ...)

MAKE_ERROR was simply used in an invalid way. Scala doesn't understand intermediate returns in a function unless we want to totally rethink all the syntax with special macros, but don't see the need of going back to Java/Fortran constructs right now.

I am quite scared by the karedos transactions I have seen in the code. Doing transactions like this NEEDS locking the two tables otherwise we really have a big risk of inconsistencies and losing transactions. Having worked for banks and finance companies I'm very aware about this. Will try to offer the simplest possibe solution to this task.

Tests are runnable from eclipse as well, but @runwith is needed by intellij as well. (Tests are now quite simplified and they can be run from the enclosed directories or simply by sbt test).